Buckaroo (old) : Add check on empty shipping phone number
This commit fixes an issue where the extension assumes the shipping phonenumber is never empty.
However there are case, in which an empty shipping phonenumber is saved.
In this case Buckaroo just copies the empty value, and adds is to to the SOAP reuest.
The SOAP request will then fail, because an empty value is sent for the shipping phone number.