[BUGFIX] Added ScopeInterface to getting configuration values
Hi Buckaroo,

Issue: when the global value of `tax/calculation/price_includes_tax` (and or `tax/calculation/shipping_includes_tax`) is different than the one on storeview level, Afterpay can give errors if there is a discount involved. This is due to the fact that the configuration value of these two was only retrieved on global value (it was missing ScopeInterface). By adding ScopeInterface::SCOPE_STORE, the value will be retrieved on the level of correct storeview.

How to reproduce:
-> Set Global value of `tax/calculation/price_includes_tax` and `tax/calculation/shipping_includes_tax` to 'Excluding Tax'
-> Set Storeview value of `tax/calculation/price_includes_tax` and `tax/calculation/shipping_includes_tax` to 'Including Tax`
-> Try to place an order with discount