Remove unneeded code in tests
This PR removes the check for a `functions.php` file which is not needed in Magento 2. Unit tests should be run similarly like `vendor/bin/phpunit --configuration dev/tests/unit/phpunit.xml.dist vendor/buckaroo/magento2/Test/Unit`. This way, the PHP settings should also be moved to the PHPUnit configuration file and not be part of the tests.